+++ darts-superresolution/src/darts_superresolution/__init__.py
@@ -1,3 +1,3 @@
 """Image superresolution of Sentinel 2 imagery for the DARTS dataset."""
 
-from darts_superresolution.upscale import Sentinel2Upscaler as Sentinel2Upscaler
+from darts_superresolution.util.upscale import Sentinel2Upscaler as Sentinel2Upscaler
