+++ darts-superresolution/src/darts_superresolution/util/patching.py
@@ -8,7 +8,7 @@ import numpy as np
 import tifffile
 import tqdm as tqdm
 import warnings
-from darts_superresolution.data_processing.util import transform_augment_tensor
+from darts_superresolution.util.util import transform_augment_tensor
 
 logger = logging.getLogger(__name__)
